Docker settings

+++ src/main/java/com/sit/student/repository/ClassesRepository.java
@@ -1,13 +1,12 @@
 package com.sit.student.repository;
 
 import com.sit.student.model.Classes;
-import com.sit.student.model.Discipline;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface ClassesRepository extends JpaRepository<Classes, Long> {
+public interface ClassesRepository extends JpaRepository<Classes, Integer> {
     List<Classes> getClassesByDisciplineId(int id);
 }
